Move block parsing functions into their own JS module so that they can be imported in node land when running Playwright tests
